Graphical Social Networking
This project is a Social Networking Website where communication is established through graphics. This project, however, is still under development. So I have included two folders for it. One folder contains the project proposal, and another folder contains the draw functionality that is later implemented into the site. Once the website is fully developed, I will include a link to that website.